moved loading and storing functions out of the store
